backend/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
requests==2.32.3
PyYAML==6.0.2
beautifulsoup4==4.13.3